requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
paho-mqtt>=1.6.1
streamlit>=1.28.0
plotly>=5.17.0
joblib>=1.3.0
